split off e2e pkg
