+++ e2e/main_other_test.go
@@ -2,8 +2,25 @@
 
 package e2e
 
-import "fmt"
+import (
+	"fmt"
 
-func mainSetup() error {
-	return fmt.Errorf("e2e tests are not supported on non-Linux platforms")
+	"github.com/qpoint-io/qtap/internal/tap"
+	"github.com/qpoint-io/qtap/pkg/connection"
+	ebpfProcess "github.com/qpoint-io/qtap/pkg/ebpf/process"
+	"github.com/qpoint-io/qtap/pkg/ebpf/socket"
+	"github.com/qpoint-io/qtap/pkg/ebpf/tls"
+	"go.uber.org/zap"
+)
+
+func NewEbpfProcManager(logger *zap.Logger, objs *tap.TapObjects) (*ebpfProcess.Manager, error) {
+	return nil, fmt.Errorf("not supported on this platform")
+}
+
+func NewEbpfSockManager(logger *zap.Logger, connMan *connection.Manager, objs *tap.TapObjects) (*socket.SocketEventManager, error) {
+	return nil, fmt.Errorf("not supported on this platform")
+}
+
+func InitTLSProbes(logger *zap.Logger, tlsProbesStr string, objs *tap.TapObjects) (*tls.TlsManager, error) {
+	return nil, fmt.Errorf("not supported on this platform")
 }
